requirements.txt
discord.py >= 1.5.0
flask
pynacl